# codec: remove unnecessary allocations in ProtoCodec.MarshalBinaryLengthPrefixed