--- a/codec/proto_codec_test.go
+++ b/codec/proto_codec_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -192,4 +193,46 @@ func TestProtoCodecUnmarshalBinaryLengthPrefixedChecks(t *testing.T) {
 
 		require.Panics(t, func() { cdc.MustUnmarshalBinaryLengthPrefixed(crafted, recv) })
 	})
+}
+
+func mustAny(msg proto.Message) *types.Any {
+	any, err := types.NewAnyWithValue(msg)
+	if err != nil {
+		panic(err)
+	}
+	return any
+}
+
+func BenchmarkProtoCodecMarshalBinaryLengthPrefixed(b *testing.B) {
+	var pCdc = codec.NewProtoCodec(types.NewInterfaceRegistry()).(*codec.ProtoCodec)
+	var msg = &testdata.HasAnimal{
+		X: 1000,
+		Animal: mustAny(&testdata.HasAnimal{
+			X: 2000,
+			Animal: mustAny(&testdata.HasAnimal{
+				X: 3000,
+				Animal: mustAny(&testdata.HasAnimal{
+					X: 4000,
+					Animal: mustAny(&testdata.HasAnimal{
+						X: 5000,
+						Animal: mustAny(&testdata.Cat{
+							Moniker: "Garfield",
+							Lives:   6,
+						}),
+					}),
+				}),
+			}),
+		}),
+	}
+
+	b.ResetTimer()
+	b.ReportAllocs()
+
+	for i := 0; i < b.N; i++ {
+		blob, err := pCdc.MarshalBinaryLengthPrefixed(msg)
+		if err != nil {
+			b.Fatal(err)
+		}
+		b.SetBytes(int64(len(blob)))
+	}
 }